Pulls out AbstractNonBlockingSocketWriter so it can be shared between client and server instead of server needing to copy/paste the socketwriter logic. It contains some complex parts like scheduling that is better not to get copy pasted.
